Move editFieldsSerivce into services/.../metadata/fields folder, change name to edit_service, change vars as needed
